removed the redundant methods
