+++ functions/src/firestoreDoc.d.ts
@@ -1,0 +1,6 @@
+/**
+ * Listing down firestore documents used in this service
+ */
+
+// Experience document data
+type Services = Record<string, {domain:string, apiKey:string, apiSecret: string, endpoints: string[]}>
